Restructured folders. Added some templates. Rudimentary server processes up/
