+++ public/404.html
@@ -1,20 +1,20 @@
-<!DOCTYPE html>
+<!<!DOCTYPE html>
 <html>
 
 <head>
 	<meta charset="utf-8">
 	<title>FaceIT</title>
-	<link rel="shortcut icon" href="/favicon.png">
+	<link rel="shortcut icon" href="favicon.png">
 
 	<link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
 	<!-- BODY FONT -->
 	<link href="https://fonts.googleapis.com/css?family=Open+Sans+Condensed:300" rel="stylesheet">
 	<!-- NAV FONT -->
-	<link rel="stylesheet" href="/style.css">
+	<link rel="stylesheet" href="style.css">
 </head>
 
 <body>
-	{{{body}}}
+	<p>"This is a 404 test page"</p>
 </body>
 
 </html>
