+++ views/homePage.handlebars
@@ -1,0 +1,1 @@
+{{> header}}
